// src/claim/dtos/claimDocument.dto.ts
export class ClaimDocument {
    claimDocumentId?: number;
    claimId?: number;
    documentName?: string;
    uniqeDocumentName?: string;
    documentTypeId?: number;
    description?: string;
    statusId?: number;
    createdDate?: Date;
    createdBy?: string;
    uplodedDate?: Date;
    uplodedBy?: string;
}